+++ edge-agent/src/Microsoft.Azure.Devices.Edge.Agent.Core/logs/ModuleLogMessage.cs
@@ -3,7 +3,6 @@ namespace Microsoft.Azure.Devices.Edge.Agent.Core.Logs
 {
     extern alias akka;
     using System;
-    using akka::Akka.IO;
     using Microsoft.Azure.Devices.Edge.Util;
     using Microsoft.Azure.Devices.Edge.Util.Json;
     using Newtonsoft.Json;
@@ -55,31 +54,5 @@ namespace Microsoft.Azure.Devices.Edge.Agent.Core.Logs
         [JsonProperty(PropertyName = "timestamp")]
         [JsonConverter(typeof(OptionConverter<DateTime>))]
         public Option<DateTime> TimeStamp { get; }
-    }
-
-
-    public class ModuleLogMessageData : ModuleLogMessage
-    {
-        public ModuleLogMessageData(
-            string iotHub,
-            string deviceId,
-            string moduleId,
-            string stream,
-            int logLevel,
-            Option<DateTime> timeStamp,
-            string text,
-            ByteString fullFrame,
-            string fullText)
-            : base(iotHub, deviceId, moduleId, stream, logLevel, timeStamp, text)
-        {
-            this.FullText = Preconditions.CheckNonWhiteSpace(fullText, fullText);
-            this.FullFrame = Preconditions.CheckNotNull(fullFrame, nameof(fullFrame));
-        }
-
-        [JsonIgnore]
-        public ByteString FullFrame { get; }
-
-        [JsonIgnore]
-        public string FullText { get; }
     }
 }
